+++ front/admin-app/src/main.ts
@@ -2,6 +2,7 @@ import { createApp } from 'vue'
 import { createVuetify } from 'vuetify'
 import { aliases, mdi } from 'vuetify/iconsets/mdi'
 import 'vuetify/styles'
+import '@mdi/font/css/materialdesignicons.css'
 
 import App from './App.vue'
 import router from './router'